Move VariableSet and OptimizableFunctorBase
