+++ src/QMCWaveFunctions/Jastrow/codegen/UserFunctor.h.in
@@ -21,7 +21,7 @@ $dire_codegen_warning
  */
 #ifndef QMCPLUSPLUS_USERFUNCTOR_H
 #define QMCPLUSPLUS_USERFUNCTOR_H
-#include "Numerics/OptimizableFunctorBase.h"
+#include "OptimizableFunctorBase.h"
 #include "OhmmsData/AttributeSet.h"
 #include <cmath>
 // #include <vector>
