+++ src/QMCWaveFunctions/VariableSet.cpp
