+++ src/QMCWaveFunctions/VariableSet.h
